chore(refactor): extract PublicationButton in individual components
