+++ src/components/item/publish/ItemPublishTab.tsx
@@ -16,7 +16,6 @@ import CoEditorsContainer from '@/components/item/publish/CoEditorsContainer';
 import EditItemDescription from '@/components/item/publish/EditItemDescription';
 import LanguagesContainer from '@/components/item/publish/LanguagesContainer';
 import LicenseContainer from '@/components/item/publish/LicenseContainer';
-import PublicationButton from '@/components/item/publish/PublicationButton';
 import PublicationStatusComponent from '@/components/item/publish/PublicationStatusComponent';
 import PublicationThumbnail from '@/components/item/publish/PublicationThumbnail';
 import { OutletType } from '@/components/pages/item/type';
@@ -26,6 +25,7 @@ import { SomeBreakPoints } from '@/types/breakpoint';
 
 import EditItemName from './EditItemName';
 import CustomizedTags from './customizedTags/CustomizedTags';
+import usePublicationButton from './publicationButtons/PublicationButton.hook';
 
 type StackOrder = { order?: number | SomeBreakPoints<number> };
 
@@ -37,6 +37,7 @@ const ItemPublishTab = (): JSX.Element => {
   const { status } = useDataSyncContext();
 
   const [notifyCoEditors, setNotifyCoEditors] = useState<boolean>(false);
+  const { publicationButton } = usePublicationButton({ item, notifyCoEditors });
 
   if (isMemberLoading) {
     return <Loader />;
@@ -110,7 +111,7 @@ const ItemPublishTab = (): JSX.Element => {
         notifyCoEditors={notifyCoEditors}
         onNotificationChanged={(enabled) => setNotifyCoEditors(enabled)}
       />
-      <PublicationButton item={item} notifyCoEditors={notifyCoEditors} />
+      {publicationButton}
     </Stack>
   );
