+++ src/components/item/publish/publicationButtons/PendingButton.tsx
@@ -1,0 +1,20 @@
+import { useBuilderTranslation } from '@/config/i18n';
+import { BUILDER } from '@/langs/constants';
+
+import PublicationButton from './PublicationButton';
+
+export const PendingButton = (): JSX.Element => {
+  const { t } = useBuilderTranslation();
+
+  return (
+    <PublicationButton
+      isLoading={false}
+      description={t(
+        BUILDER.LIBRARY_SETTINGS_VALIDATION_STATUS_PENDING_AUTOMATIC,
+      )}
+      elements={[]}
+    />
+  );
+};
+
+export default PendingButton;
